fix: type colomn afisha_id, place icon.

+++ database/migrations/2024_06_25_114232_add_colomn_afisha_id_in_table_events.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->integer('afisha7_id')->nullable();
+            $table->bigInteger('afisha7_id')->nullable();
         });
     }
